pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ena-mcp-server"
version = "0.1.0"
description = "MCP server exposing ENA (European Nucleotide Archive) REST APIs to AI agents"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "Prachi", email = "prachi@example.com" },
]
keywords = ["mcp", "ena", "bioinformatics", "llm", "genomics", "rest-api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "cachetools>=5.3.3",
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.1",
    "anyio>=4.4.0",
    "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
    "coverage>=7.5.1",
]

[project.scripts]
ena-mcp = "ena_mcp.server:main"

[project.urls]
Homepage = "https://github.com/enasequence/ena-mcp-server"
Documentation = "https://github.com/enasequence/ena-mcp-server/blob/main/docs/architecture.md"
Repository = "https://github.com/enasequence/ena-mcp-server"
Issues = "https://github.com/enasequence/ena-mcp-server/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ena_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
mypy_path = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
